+++ rpc/lib/client/http_json_client_test.go
