+++ rpc/lib/client/ws_client_test.go
@@ -35,9 +35,15 @@ func (h *myHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	}
 	defer conn.Close() // nolint: errcheck
 	for {
-		messageType, _, err := conn.ReadMessage()
+		messageType, in, err := conn.ReadMessage()
 		if err != nil {
 			return
+		}
+
+		var req types.RPCRequest
+		err = json.Unmarshal(in, &req)
+		if err != nil {
+			panic(err)
 		}
 
 		h.mtx.RLock()
@@ -49,7 +55,7 @@ func (h *myHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		h.mtx.RUnlock()
 
 		res := json.RawMessage(`{}`)
-		emptyRespBytes, _ := json.Marshal(types.RPCResponse{Result: res})
+		emptyRespBytes, _ := json.Marshal(types.RPCResponse{Result: res, ID: req.ID})
 		if err := conn.WriteMessage(messageType, emptyRespBytes); err != nil {
 			return
 		}
