+++ rpc/lib/types/types_test.go
@@ -41,12 +41,9 @@ func TestResponses(t *testing.T) {
 		s := fmt.Sprintf(`{"jsonrpc":"2.0","id":%v,"result":{"Value":"hello"}}`, tt.expected)
 		assert.Equal(s, string(b))
 
-		d := RPCParseError(jsonid, errors.New("Hello world"))
+		d := RPCParseError(errors.New("Hello world"))
 		e, _ := json.Marshal(d)
-		f := fmt.Sprintf(
-			`{"jsonrpc":"2.0","id":%v,"error":{"code":-32700,"message":"Parse error. Invalid JSON","data":"Hello world"}}`,
-			tt.expected,
-		)
+		f := `{"jsonrpc":"2.0","error":{"code":-32700,"message":"Parse error. Invalid JSON","data":"Hello world"}}`
 		assert.Equal(f, string(e))
 
 		g := RPCMethodNotFoundError(jsonid)
